+++ lib/td_dq/search/aggregations.ex
@@ -62,11 +62,20 @@ defmodule TdDq.Search.Aggregations do
     |> Enum.map(&content_term(&1, scope))
   end
 
+  defp filter_content_term(%{"type" => "system"}), do: true
   defp filter_content_term(%{"values" => values}) when is_map(values), do: true
   defp filter_content_term(_), do: false
 
   defp content_term(%{"name" => field, "type" => "user"}, "dq") do
     {field, %{terms: %{field: "df_content.#{field}.raw", size: 50}}}
+  end
+
+  defp content_term(%{"name" => field, "type" => "system"}, "dq") do
+    {field,
+     %{
+       nested: %{path: "df_content.#{field}"},
+       aggs: %{distinct_search: %{terms: %{field: "df_content.#{field}.external_id.raw", size: 50}}}
+     }}
   end
 
   defp content_term(%{"name" => field}, "dq") do