+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 
 - [TD-3047] Migrate executions functionality from rules to implementations
 
+## [4.6.0] 2020-10-19
+
+### Added
+
+- [TD-2485]:
+  - Enrich template fields from cache
+  - Mappings for system type of templates
+  
 ## [4.5.1] 2020-10-06
 
 ### Fixed